requirements.txt
fastapi
uvicorn
gunicorn
jinja2
fastapi-cache2
httpx
sqlmodel
markdown
beautifulsoup4
sqlitecloud
sqlalchemy-sqlitecloud
smartypants
cloudinary
asyncer
polars
great-tables
pytz
atproto
APScheduler